+++ src-tauri/src/history.rs
@@ -1509,6 +1509,7 @@ mod tests {
             artist: "Artist".to_owned(),
             album: "Album".to_owned(),
             release_year: Some(2024),
+            original_year: Some(2024),
             rating: None,
             loved: false,
             love_state: LoveState::Neutral,